+++ front/app/containers/Admin/index.tsx
@@ -20,6 +20,8 @@ import { endsWith } from 'utils/helperUtils';
 import 'assets/semantic/semantic.min.css';
 import { rgba } from 'polished';
 
+import Outlet from 'components/Outlet';
+
 const Container = styled.div`
   display: flex;
   background: ${colors.background};
@@ -52,7 +54,7 @@ const Container = styled.div`
   }
 `;
 
-const RightColumn = styled.div`
+export const RightColumn = styled.div`
   flex-grow: 1;
   flex-shrink: 1;
   flex-basis: 0;
@@ -117,6 +119,8 @@ const AdminPage = memo<Props & WithRouterProps>(
     const [adminFullWidth, setAdminFullWidth] = useState(false);
     const [adminNoPadding, setAdminNoPadding] = useState(false);
 
+    const [adminFullWidthContent, setAdminFullWidthContent] = useState(false);
+
     useEffect(() => {
       const subscriptions = [
         globalState
@@ -130,6 +134,9 @@ const AdminPage = memo<Props & WithRouterProps>(
         subscriptions.forEach((subscription) => subscription.unsubscribe());
       };
     }, []);
+
+    const setAdminFullWidthContentToVisible = (isVisible) =>
+      setAdminFullWidthContent(isVisible);
 
     const userCanViewAdmin = () =>
       hasPermission({
@@ -153,19 +160,15 @@ const AdminPage = memo<Props & WithRouterProps>(
     const noPadding =
       adminNoPadding ||
       pathname.includes('admin/dashboard') ||
-      pathname.includes('admin/processing') ||
       pathname.includes('admin/insights');
 
     const fullWidth =
       adminFullWidth === true ||
       endsWith(pathname, 'admin/moderation') ||
       pathname.includes('admin/dashboard') ||
-      pathname.includes('admin/processing') ||
       pathname.includes('admin/insights');
 
-    const whiteBg =
-      endsWith(pathname, 'admin/moderation') ||
-      pathname.includes('admin/processing');
+    const whiteBg = endsWith(pathname, 'admin/moderation');
 
     return (
       <HasPermission
@@ -174,14 +177,23 @@ const AdminPage = memo<Props & WithRouterProps>(
       >
         <ThemeProvider theme={chartTheme}>
           <Container className={`${className} ${whiteBg ? 'whiteBg' : ''}`}>
-            <Sidebar />
-            <RightColumn
-              className={`${fullWidth && 'fullWidth'} ${
-                noPadding && 'noPadding'
-              }`}
-            >
-              {children}
-            </RightColumn>
+            {!adminFullWidthContent && (
+              <>
+                <Sidebar />
+                <RightColumn
+                  className={`${fullWidth && 'fullWidth'} ${
+                    noPadding && 'noPadding'
+                  }`}
+                >
+                  {children}
+                </RightColumn>
+              </>
+            )}
+            <Outlet
+              id="app.containers.Admin.contentBuilderLayout"
+              onMount={setAdminFullWidthContentToVisible}
+              childrenToRender={children}
+            />
           </Container>
         </ThemeProvider>
       </HasPermission>