+++ front/app/resources/GetTopics.tsx
@@ -4,20 +4,12 @@ import { Subscription, BehaviorSubject, of, combineLatest } from 'rxjs';
 import { distinctUntilChanged, switchMap, map } from 'rxjs/operators';
 import {
   ITopicData,
+  ITopicsQueryParams,
   topicByIdStream,
   topicsStream,
-  Code,
 } from 'services/topics';
 import { isNilOrError, NilOrError, reduceErrors } from 'utils/helperUtils';
-
-interface InputProps {
-  // Don't use the ids and the query parameters (code, exclude_code, sort) together
-  // Only one of the two at a time.
-  topicIds?: string[];
-  code?: Code;
-  exclude_code?: Code;
-  sort?: 'new' | 'custom';
-}
+import { Parameters as InputProps } from 'hooks/useTopics';
 
 type children = (renderProps: GetTopicsChildProps) => JSX.Element | null;
 
@@ -47,54 +39,63 @@ export default class GetTopics extends React.Component<Props, State> {
   }
 
   componentDidMount() {
-    const { topicIds, code, exclude_code, sort } = this.props;
+    const { topicIds, code, excludeCode, sort, forHomepageFilter } = this.props;
 
     this.inputProps$ = new BehaviorSubject({
       topicIds,
       code,
-      exclude_code,
+      excludeCode,
       sort,
+      forHomepageFilter,
     });
 
     this.subscriptions = [
       this.inputProps$
         .pipe(
           distinctUntilChanged((prev, next) => isEqual(prev, next)),
-          switchMap(({ topicIds, code, exclude_code, sort }) => {
-            const queryParameters = { code, exclude_code, sort };
+          switchMap(
+            ({ topicIds, code, excludeCode, sort, forHomepageFilter }) => {
+              if (topicIds) {
+                if (topicIds.length > 0) {
+                  return combineLatest(
+                    topicIds.map((id) => {
+                      return topicByIdStream(id).observable.pipe(
+                        map((topic) =>
+                          !isNilOrError(topic) ? topic.data : topic
+                        )
+                      );
+                    })
+                  );
+                }
 
-            if (topicIds) {
-              if (topicIds.length > 0) {
-                return combineLatest(
-                  topicIds.map((id) => {
-                    return topicByIdStream(id).observable.pipe(
-                      map((topic) =>
-                        !isNilOrError(topic) ? topic.data : topic
-                      )
-                    );
-                  })
+                return of(null);
+              } else {
+                const queryParameters: ITopicsQueryParams = {
+                  code,
+                  exclude_code: excludeCode,
+                  sort,
+                  for_homepage_filter: forHomepageFilter,
+                };
+
+                return topicsStream({ queryParameters }).observable.pipe(
+                  map((topics) => topics.data)
                 );
               }
-
-              return of(null);
-            } else {
-              return topicsStream({ queryParameters }).observable.pipe(
-                map((topics) => topics.data)
-              );
             }
-          })
+          )
         )
         .subscribe(reduceErrors<ITopicData>(this.setTopics.bind(this))),
     ];
   }
 
   componentDidUpdate() {
-    const { topicIds, code, exclude_code, sort } = this.props;
+    const { topicIds, code, excludeCode, sort, forHomepageFilter } = this.props;
     this.inputProps$.next({
       topicIds,
       code,
-      exclude_code,
+      excludeCode,
       sort,
+      forHomepageFilter,
     });
   }